ref: modularise views.py
